requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
distlib==0.3.6
filelock==3.8.0
idna==3.4
platformdirs==2.5.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
slackclient==1.3.2
urllib3==1.26.12
virtualenv==20.16.5
websocket-client==0.54.0
slack_sdk
slack_bolt
slackeventsapi